ortho: use a 'size_t' for number of vertices, 'nvs'
This pays down some cast noise and other debt from prior commits.

+++ lib/ortho/rawgraph.h
@@ -21,7 +21,7 @@ typedef struct {
 } vertex;
 
 typedef struct {
-  int nvs;
+  size_t nvs;
   vertex* vertices;
 } rawgraph;
